# dist/api-docs.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a API documentation for OB server.  You can find out more about Ocean Builders at [https://ocean.builders/](https://ocean.builders/)",
    "version": "0.7",
    "title": "Ocean Builder",
    "contact": {
      "email": "diego@scissorboy.com"
    }
  },
  "host": "oceanbuilders.herokuapp.com",
  "basePath": "/v1/api/",
...
  "schemes": [
    "https"
  ],
  "paths": {
    "/auth": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Create new user associated with new seapod",
        "description": "",
        "operationId": "SignUpWithSeaPodCreation",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "User Information and Seapod Information",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "user": {
                  "type": "object",
                  "properties": {
                    "firstName": {
                      "type": "string"
                    },
                    "lastName": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string",
                      "format": "email"
                    },
                    "mobileNumber": {
                      "type": "string",
                      "pattern": "^(\\+\\d{1,3}[- ]?)?\\d{10}$"
                    },
                    "password": {
                      "type": "string",
                      "pattern": "^.*(?=.{8,})((?=.*[!@#$%^&*()\\-_=+{};:,<.>]){1})(?=.*\\d)((?=.*[a-z]){1})((?=.*[A-Z]){1}).*$"
                    },
                    "country": {
                      "type": "string"
                    }
                  }
                },
                "seaPod": {
                  "type": "object",
                  "properties": {
                    "SeaPodName": {
                      "type": "string"
                    },
                    "exteriorFinish": {
                      "type": "string"
                    },
                    "exterioirColor": {
                      "type": "string",
                      "pattern": "^0xFF[A-F0-9]{6}$"
                    },
                    "sparFinish": {
                      "type": "string"
                    },
                    "sparDesign": {
                      "type": "string"
                    },
                    "deckEnclosure": {
                      "type": "string"
                    },
                    "bedAndLivingRoomEnclousure": {
                      "type": "string"
                    },
                    "power": {
                      "type": "string"
                    },
                    "powerUtilities": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "underWaterRoomFinishing": {
                      "type": "string"
                    },
                    "underWaterWindows": {
                      "type": "string"
                    },
                    "soundSystem": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "masterBedroomFloorFinishing": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "masterBedroomInteriorWallColor": {
                      "type": "string",
                      "pattern": "^0xFF[A-F0-9]{6}$"
                    },
                    "livingRoomloorFinishing": {
                      "type": "string"
                    },
                    "livingRoomInteriorWallColor": {
                      "type": "string",
                      "pattern": "^0xFF[A-F0-9]{6}$"
                    },
                    "kitchenfloorFinishing": {
                      "type": "string"
                    },
                    "kitchenInteriorWallColor": {
                      "type": "string",
                      "pattern": "^0xFF[A-F0-9]{6}$"
                    },
                    "hasWeatherStation": {
                      "type": "boolean"
                    },
                    "entryStairs": {
                      "type": "boolean"
                    },
                    "hasFathometer": {
                      "type": "boolean"
                    },
                    "hasCleanWaterLevelIndicator": {
                      "type": "boolean"
                    },
                    "interiorBedroomWallColor": {
                      "type": "string",
                      "pattern": "^0xFF[A-F0-9]{6}$"
                    },
                    "deckFloorFinishMaterial": {
                      "type": "string"
                    },
                    "seaPodStatus": {
                      "type": "string"
                    }
                  }
                }
              },
              "example": {
                "user": {
                  "firstName": "Name",
                  "lastName": "Name",
                  "email": "user@example.com",
                  "mobileNumber": "+201564324522",
                  "password": "String@14*",
                  "country": "Spain"
                },
                "seaPod": {
                  "SeaPodName": "Sea Breeze",
                  "exteriorFinish": "exfinish",
                  "exterioirColor": "0xff123456",
                  "sparFinish": "sparFinish",
                  "sparDesign": "sparDesign",
                  "deckEnclosure": "deckEnclosure",
                  "bedAndLivingRoomEnclousure": "bedAndLivingRoomEnclousure",
                  "power": "power",
                  "powerUtilities": [],
                  "underWaterRoomFinishing": "underWaterRoomFinishing",
                  "underWaterWindows": "underWaterWindows",
                  "soundSystem": [],
                  "masterBedroomFloorFinishing": [],
                  "masterBedroomInteriorWallColor": "0xff123456",
                  "livingRoomloorFinishing": "livingRoomloorFinishing",
                  "livingRoomInteriorWallColor": "0xff123456",
                  "kitchenfloorFinishing": "kitchenfloorFinishing",
                  "kitchenInteriorWallColor": "0xff123456",
                  "hasWeatherStation": false,
                  "entryStairs": true,
                  "hasFathometer": false,
                  "hasCleanWaterLevelIndicator": false,
                  "interiorBedroomWallColor": "0xff123456",
                  "deckFloorFinishMaterial": "deckFloorFinishMaterial",
                  "seaPodStatus": "seaPodStatus"
                }
              }
            }
          },
          {
            "in": "header",
            "name": "notificationToken",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "hardwareId",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "model",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "user and sea pod data are required, Validation Error"
          },
          "500": {
            "description": "Internal Server Error, Header Parameters is missing"
          }
        }
      },
      "put": {
        "tags": [
          "Auth"
        ],
        "summary": "Login with email and password",
        "description": "",
        "operationId": "login",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "User Information and Seapod Information",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "format": "email"
                },
                "password": {
                  "type": "string",
                  "pattern": "^.*(?=.{8,})((?=.*[!@#$%^&*()\\-_=+{};:,<.>]){1})(?=.*\\d)((?=.*[a-z]){1})((?=.*[A-Z]){1}).*$"
                }
              },
              "example": {
                "email": "user@example.com",
                "password": "String@14*"
              }
            }
          },
          {
            "in": "header",
            "name": "notificationToken",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "hardwareId",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "model",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Your email and password combination was incorrect. Please try again."
          },
          "500": {
            "description": "Internal Server Error, Header Parameters is missing"
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Login with token",
        "description": "",
        "operationId": "autoLogin",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "x-auth-token",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "hardwareId",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Your email and password combination was incorrect. Please try again."
          },
          "500": {
            "description": "Internal Server Error, Header Parameters is missing"
          }
        }
      }
    },
...
  }
}